refactor: split dashboard code
