+++ src/components/chart/style.css
@@ -1,0 +1,6 @@
+.chart {
+    /* background-color:tomato; */
+    width: 60%;
+    height: 80vh;
+    padding-left: 20px;
+}
